requirements.txt
Flask
numpy
scikit-learn
joblibs